wip: create vault

+++ frontend/src/constants/ABI/PartnerContractsDeployer.ts
@@ -127,9 +127,9 @@ export const ABI = [
         type: "uint8",
       },
       {
-        internalType: "uint8",
+        internalType: "uint256",
         name: "_maxAmtPercentInRp",
-        type: "uint8",
+        type: "uint256",
       },
     ],
     name: "registerAsPartner",
